requirements.txt
python-telegram-bot[job-queue]==21.5
python-dotenv==1.0.0
flask==3.0.3
requests==2.31.0
aiohttp==3.9.1